SELC-3480 Add modal for spid login selection
